衝突判定が働かないバグとsuperWallのバグを修正

+++ Balls/Collider.cpp
@@ -104,7 +104,7 @@ bool Collider::CollisionBallWall(SphereColliderComponent * sphere, BoxColliderCo
 	BoxInfo boxInfo = box->GetIndo();
 
 	//ìØÇ∂ç¿ïWÇ…Ç¢ÇÈÇ©
-	if (boxInfo.position.x == sphereInfo.position.x)
+	if ((int)boxInfo.position.x == (int)sphereInfo.position.x)
 	{
 		return true;
 	}
